Comentarii

+++ PlatformaSocialBookmarking/Controllers/CommentsController.cs
@@ -1,7 +1,11 @@
-﻿using Microsoft.AspNetCore.Identity;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using PlatformaSocialBookmarking.Data;
 using PlatformaSocialBookmarking.Models;
+using Microsoft.EntityFrameworkCore;
+
 
 namespace PlatformaSocialBookmarking.Controllers
 {
@@ -9,9 +13,21 @@ namespace PlatformaSocialBookmarking.Controllers
     {
         private readonly ApplicationDbContext db;
 
-        public CommentsController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        private readonly RoleManager<IdentityRole> _roleManager;
+
+        public CommentsController(
+            ApplicationDbContext context,
+            UserManager<ApplicationUser> userManager,
+            RoleManager<IdentityRole> roleManager
+            )
         {
             db = context;
+            
+            _userManager = userManager;
+
+            _roleManager = roleManager;
         }
 
 
@@ -32,36 +48,69 @@ namespace PlatformaSocialBookmarking.Controllers
         }
 
         [HttpPost]
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Delete(int id)
         {
             Comment comm = db.Comments.Find(id);
-            db.Comments.Remove(comm);
-            db.SaveChanges();
-            return Redirect("/Bookmarks/Show" + comm.BookmarkId);
+
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                db.Comments.Remove(comm);
+                db.SaveChanges();
+                return Redirect("/Bookmarks/Show/" + comm.BookmarkId);
+            }
+
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa stergeti comentariul";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Bookmarks");
+            }
         }
 
-
+        [Authorize(Roles = "User,Editor,Admin")]
         public IActionResult Edit(int id)
         {
             Comment comm = db.Comments.Find(id);
 
-            return View(comm);
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
+            {
+                return View(comm);
+            }
+
+            else
+            {
+                TempData["message"] = "Nu aveti dreptul sa editati comentariul";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Bookmarks");
+            }
         }
 
         [HttpPost]
+        [Authorize(Roles = "User, Editor, Admin")]
         public IActionResult Edit(int id, Comment requestComment)
         {
             Comment comm = db.Comments.Find(id);
 
-            if (ModelState.IsValid)
+            if (comm.UserId == _userManager.GetUserId(User) || User.IsInRole("Admin"))
             {
-                comm.Content = requestComment.Content;
-                db.SaveChanges();
-                return Redirect("/Bookmarks/Show" + comm.BookmarkId);
+
+                if (ModelState.IsValid)
+                {
+                    comm.Content = requestComment.Content;
+                    db.SaveChanges();
+                    return Redirect("/Bookmarks/Show/" + comm.BookmarkId);
+                }
+                else
+                {
+                    return View(requestComment);
+                }
             }
             else
             {
-                return View(requestComment);
+                TempData["message"] = "Nu aveti dreptul sa faceti modificari";
+                TempData["messageType"] = "alert-danger";
+                return RedirectToAction("Index", "Bookmarks");
             }
         }
     }
